+++ src/channels/Channel.ts
