+++ src/index.ts
@@ -2,5 +2,5 @@ export * from './agents/index.js'
 export * from './encrypt/index.js'
 export * from './sign/index.js'
 export * from './derive/index.js'
-export * from './channel/index.js'
+export * from './channels/index.js'
 export * from './hash/index.js'
